Print models from db in Dashboard page
The Dashboard page shows a table populated with the information about
models stored in the database. A new endpoint has been added to
support this feature. The endpoint is identical to one used by the
configurator. We may consider removing merging the two endpoint if
they end up being the same.