Use new json converter from clafer BESSPIN
The json format from the Clafer BESSPIN tool simplifies the data model
and result in simplification in both the client and the server. The
variety of clafer cardinalities and group cardinalities have been
simplified, it leads to simplifications of the code on the client
side. On the server side, most of the code has been deleted and what
remains has been moved to a dedicated `configurator_shim.py` module.