Compute configured feature model in .fm.json format
This is done on the server side by combining the fm.json feature model
and the configurations (feature selections). Every feature selection
is added to the list of constraints. This commit also cleanup some
deadcode related to naming feature selection based on path from
feature to root of the feature model. This is less with the fm.json
format.