skip coverage metrics for the modules we will never test
the index file for the UI has no testable functionality. The serviceWorkers _may_ need testing in the future if we decide to use it, but I believe we will actually remove it instead. Sagas should be simple coordinators and all the tests I've seen for sagas are overly complex for the value they provide. App.tsx is similar to index.tsx in that it is simply a configuration of the application