update tests with correct type information
It appears that create-react-app made a design choice not to do any type-checking when running unit tests (https://stackoverflow.com/questions/55938211/typescript-errors-doesnt-show-when-running-tests-through-jest-with-create-react)

I only noticed these errors via the linter. I have added the type-checking to the `npm test` command, but it only will test when starting up the test runner since it runs in watch-mode. Not ideal, but better than nothing.