WIP: first pass at a "workflow" UX
This is a fairly big commit focused on suggesting a potential UX for communicating the testing workflow as succinctly as possible to the user. Rather than reading the code, this should be reviewed by starting up the UI and exploring the actual, rendered page.

Once the UX is settled on, there will be work to both clean up the code as well as implement actual state-management and the underlying API to support the data used in this page.