shift the workflow::sysconfig relationship to point from config to workflow
Realized that in the actual workflow, a user will need to associate a system-configuration-input to a given workflow rather than vice-versa, so I shifted the foreign-key relationship to originate on the SystemConfigurationInput side.

Also added the SQLAlchemy "relationship" to auto-populate the workflow with the SystemConfigurationInput data for ease of access.